Replace Google charts table with custom table
Google charts' table view was slowing down the page because of all the
event listeners, so it has been replaced with a simple custom table
view.

The current table view is very simple and it displays _all_ the data.
More features will be added in future PRs, such as ability to download
the table as a CSV, simple pagination, and ability to view table with
another viz.