Update validators to allow conditional validation.
Easiest way to handle this without a bunch of complex logic around conditional validators was to just pass in the entire model and allow the validator function to decide whether something is valid. This means that the validator function can look at another model attribute and choose to not validate, thus returning true, because of the value of another attribute. This allows us to ONLY validate an explorer's name attribute IF the explorer has a "saving" value of true (eg: only validate the name "on save").