Fix deleting of saved queries on explorer
Previously, through the UI, you were only able to delete bookmarked queries that
belonged to you and only you. However, because we're not saving user.id
(yet), all saved queries are delete-able by anyone who has access to the
project. The test for this scenario is skipped in this commit.

This commit also renames `body` to `callbackBody` on line 37 of
`KeenSavedQueries.js`, because the `body` variable is already used in
that scope.