Removing use of user object to attach to saved queries and to use to prevent CRUD actions.
We decided to remove the user objectb data from the metadata cache. We were originally using this to attach names to saved queries and to also only allow the author to save, delete, and update queries. However, this doesn't make sense given that the API doesn't work this way, so we're removing this functionality. This also means removing the "mine" vs. "team" filter. Additionally, because saved query quantities are low for all projects we're also removing the search feature for the time being. We may add it back at a later time.