Fix parsing issue
Don't return dates that come back as "Invalid date" when formatted my moment.js. Eg: If the value is "true" from a Boolean filter that's been switched to a datetime filter, that date will not properly parse. In which case, just return a date time for start of day, yesterday.

Also, remove the helper fucntion. Not needed as it's only called inside the 'Datetime' coercion function and nowhere else.